% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.calc.R
\name{IV.calc}
\alias{IV.calc}
\title{Instrumental variable analysis for multiple imputed datasets.}
\usage{
IV.calc(instrument = NULL, therapy = NULL, outcome = NULL,
  mids = NULL)
}
\arguments{
\item{instrument}{A character value of the name of the instrumental variable (any type)}

\item{therapy}{A character value of the name of the binary therapy}

\item{outcome}{A character value of the name of the outcome variable, both binary as ordinal.}

\item{mids}{A mids object, of the multiple imputed dataset which we will use for the IV analysis.}
}
\value{
Returns a vector with the effect estimate with 95% CI.
}
\description{
Instrumental variable analysis for multiple imputed datasets.
}
